backend/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cloudinary==1.29.0
Django==4.0.4
django-cors-headers==3.13.0
django-filter==21.1
djangorestframework==3.13.1
psycopg2 @ git+https://github.com/nwcell/psycopg2-windows.git@4039d982152492453abdfb72d0054230fcb42c30
psycopg2-binary==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.0.0
